refactor!: remove unnecessary error enum and result type aliases
ref: https://github.com/tauri-apps/tauri/issues/7756
