+++ .changes/tauri-extranous-result-types-removal.md
@@ -1,0 +1,6 @@
+---
+'tauri': 'major:breaking'
+---
+
+- Removed `tauri::path::Error` and added its variants to `tauri::Error`
+- Removed `tauri::plath::Result` and `tauri::plugin::Result` aliases, you should use `tauri::Result`
